update: update backend controllers database init and interface imports and exports
